los controlladores de categoria y producto solo les falta editar y remplazar la imagen

+++ database/migrations/2024_01_31_022353_create_productos.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->double('precio');
             $table->string('tamanio')->nullable(); 
             $table->string('descripcion');
-            $table->tinyInteger('disponibilidad');
+            $table->integer('disponibilidad');
             $table->string('imagen_1');
             $table->string('imagen_2');
             $table->integer('descuento')->nullable();
